Split parser expression suffixes
